refactor(tangle-cloud): consolidate request variant display into CommitmentSection
Move ModalSafeTooltip and request variant/exposure logic from ServiceRequestSummary
into CommitmentSection, removing the separate "Request Variant" section. Also update
ExposureCommitmentInput to always show the operator exposure row with a "protocol
default" label when exposure is 100%.
